gitapiのusecaseを追加
